+++ Arduino/Wheelly/Utils.h
@@ -1,0 +1,8 @@
+#ifndef Utils_h
+#define Utils_h
+
+#include "Arduino.h"
+
+float normalRad(float deg);
+
+#endif